% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc.gapbk-package.R
\docType{package}
\name{moc.gapbk-package}
\alias{moc.gapbk-package}
\title{moc.gapbk: Multi-Objective Clustering Guided by a-Priori Biological Knowledge}
\description{
The \pkg{moc.gapbk} package implements the MOC-GaPBK algorithm proposed
by Parraga-Alava and others (2018). It combines NSGA-II with
Path-Relinking and Pareto Local Search to discover clustering
solutions that are good with respect to two objective functions
simultaneously, typically defined from two distance matrices: one
over the data itself and one encoding a-priori biological knowledge.
}
\details{
The main user-facing function is \code{\link{moc.gapbk}}. The legacy
name \code{\link{moc.gabk}} is preserved as a deprecated alias for
backward compatibility.
}
\references{
J. Parraga-Alava, M. Dorn, M. Inostroza-Ponta (2018).
\emph{A multi-objective gene clustering algorithm guided by apriori
biological knowledge with intensification and diversification
strategies}. BioData Mining. 11(1) 1-16.
\doi{10.1186/s13040-018-0178-4}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jorgeklz/package-moc.gapbk}
  \item Report bugs at \url{https://github.com/jorgeklz/package-moc.gapbk/issues}
}

}
\author{
\strong{Maintainer}: Jorge Parraga-Alava \email{jorge.parraga@utm.edu.ec} (\href{https://orcid.org/0000-0001-8558-9122}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jorge Parraga-Alava \email{jorge.parraga@utm.edu.ec} (\href{https://orcid.org/0000-0001-8558-9122}{ORCID}) [copyright holder]
  \item Marcio Dorn
  \item Mario Inostroza-Ponta
}

}
\keyword{internal}
